% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\name{anderson_darling_test}
\alias{anderson_darling_test}
\title{Performs two-sample AD test}
\usage{
anderson_darling_test(x, y)
}
\arguments{
\item{x}{x celltype x}

\item{y}{y celltype y}
}
\value{
result of AD test pvalues
}
\description{
Performs two-sample AD test
}
