% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\name{cramer_von_masis}
\alias{cramer_von_masis}
\title{Performs two-sample CVM test}
\usage{
cramer_von_masis(x, y)
}
\arguments{
\item{x}{x celltype x}

\item{y}{y celltype y}
}
\value{
result of CVM test pvalues
}
\description{
Performs two-sample CVM test
}
