% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\name{distribution_test}
\alias{distribution_test}
\title{A test can be selected here between 3 test and a parametric estimation}
\usage{
distribution_test(x, y, method)
}
\arguments{
\item{x}{x celltype x}

\item{y}{y celltype y}

\item{method}{one of the tests with values 1,2,3}
}
\value{
result of CVM test pvalues
}
\description{
A test can be selected here between 3 test and a parametric estimation
}
