% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\name{falpha To be used in set_sample_function}
\alias{falpha To be used in set_sample_function}
\alias{falpha}
\title{A function for generating c}
\usage{
falpha(x, p)
}
\arguments{
\item{x}{an input data}

\item{p}{coming from params data}
}
\value{
one value
}
\description{
A function for generating c
}
