% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\name{fc To be used in set_sample_function}
\alias{fc To be used in set_sample_function}
\alias{fc}
\title{A function for generating c}
\usage{
fc(x, p)
}
\arguments{
\item{x}{an input data}

\item{p}{coming from params data}
}
\value{
a vector of values
}
\description{
A function for generating c
}
