% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\name{geometric_mean To be used in set_sample_function}
\alias{geometric_mean To be used in set_sample_function}
\alias{geometric_mean}
\title{A function finding the geometric mean of the data}
\usage{
geometric_mean(data)
}
\arguments{
\item{data}{the input data}
}
\value{
geometric mean of each celltype
}
\description{
A function finding the geometric mean of the data
}
