% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\name{get_para_bayesion}
\alias{get_para_bayesion}
\title{The retrieving the parameters by using get_para_moments}
\usage{
get_para_bayesion(p, nIter = 1000)
}
\arguments{
\item{p}{the values come from pois-beta sample}

\item{nIter}{the number of iterations}
}
\value{
returns a list of parameters for bayesian estimation
}
\description{
The retrieving the parameters by using get_para_moments
}
