% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\name{get_para_moments}
\alias{get_para_moments}
\title{The posterior parameters}
\usage{
get_para_moments(p)
}
\arguments{
\item{p}{the values come from pois-beta sampke}
}
\value{
returns a list of parameters alpha, beta , gamma and c
}
\description{
The posterior parameters
}
