% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\name{goodness_of_fit}
\alias{goodness_of_fit}
\title{performs a goodness of fit test and runs a CvM test}
\usage{
goodness_of_fit(p, params, mean_params)
}
\arguments{
\item{p}{the values come from pois-beta sample}

\item{params}{parameters from bayesian estimation}

\item{mean_params}{from the bayesian estimation}
}
\value{
returns the result of hypothesis test to check if the genes come from the same distribution
}
\description{
performs a goodness of fit test and runs a CvM test
}
