% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\name{kolmogorov_smirnov_test}
\alias{kolmogorov_smirnov_test}
\title{Performs two-sample KS test}
\usage{
kolmogorov_smirnov_test(x, y)
}
\arguments{
\item{x}{x celltype x}

\item{y}{y celltype y}
}
\value{
result of ks test pvalues
}
\description{
Performs two-sample KS test
}
