% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\name{rand_Poisson_beta}
\alias{rand_Poisson_beta}
\title{generating parametres from poisson-beta distribution}
\usage{
rand_Poisson_beta(params, n)
}
\arguments{
\item{params}{from the param data where alpha, beta, gamma are extracted}

\item{n}{the number of samples}
}
\value{
returns n samples from pois-beta distribution
}
\description{
generating parametres from poisson-beta distribution
}
