% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\name{set_sample_function}
\alias{set_sample_function}
\title{A function for parametric estimation of variabels in D3E}
\usage{
set_sample_function(func)
}
\arguments{
\item{func}{A function inherited from fc, fgamma,falpha,fbeta}
}
\value{
a list with a sample, a parameter, maen of sample
}
\description{
A function for parametric estimation of variabels in D3E
}
