% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code1.R
\name{weight_normalization gets the weights}
\alias{weight_normalization gets the weights}
\alias{weight_normalization}
\alias{normalized_data}
\title{A function to gets the weights for normalizing the data}
\usage{
weight_normalization(data)

normalized_data(data, weights)
}
\arguments{
\item{data}{normalized_data}

\item{weights}{from normalizatiom weight}
}
\value{
the normalized data

the normalized data
}
\description{
A function to gets the weights for normalizing the data

A function to normalize the data
}
